requirements.txt
# pyTelegramBotAPI
python-telegram-bot
google-cloud-firestore
firebase-admin